Pick the right slicing overload
Fixes the package build

The Xcode project is building with the 4.0 tools, but the package
itself is still using the Swift 3.0 package layout and tools.  This
means that the compiler does not see the availability annotations that
suppress init?(_ other: String) causing mismatched optionality to
occur if you build the package with SwiftPM but not with Xcode.  Because
it would be more invasive to re-layout the package for Swift 4.0, I'm
just going to force it to pick the right overload with a cast.